Fix handling of exceptions when osquery goes away (#83)
This resolves an issue where the extension does not shut down if osquery
goes away unexpectedly (without sending a shutdown signal via Thrift).

The scenario could be reliably reproduced by running `osqueryi`,
connecting an extension, and then sending a `SIGKILL` to `osqueryi`. The
exception thrown in the `start_watcher` function would be of type
`thrift.transport.TTransport.TTransportException`, and would cause the
watcher thread to exit without exiting the rest of the program.

This is a possible fix for issues that users have experienced with
extensions reconnecting after the Watchdog kills osquery.